make msvc happy

+++ apps/daesop/rentry.cpp
@@ -57,7 +57,7 @@ struct RESEntryHeader *getResourceEntryHeader(int aNumber, FILE *aResFile,
 
 	loEntryHeaderSize = sizeof(struct RESEntryHeader);
 	loEntryHeader = (struct RESEntryHeader *) malloc(loEntryHeaderSize);
-	loReadSize = fread(loEntryHeader, 1, loEntryHeaderSize, aResFile);
+	loReadSize = static_cast<unsigned int>(fread(loEntryHeader, 1, loEntryHeaderSize, aResFile));
 	if (loEntryHeaderSize != loReadSize) {
 		free(loEntryHeader);
 		printf("The resource entry header could not be read!\n");
@@ -121,7 +121,7 @@ unsigned char *readResourceBinary(int aResourceNumber, FILE *aResFile,
 		free(loResEntryHeader);
 		return (NULL);
 	}
-	loReadSize = fread(loBuffer, 1, loDataSize, aResFile);
+	loReadSize = static_cast<unsigned int>(fread(loBuffer, 1, loDataSize, aResFile));
 	if (loReadSize != loDataSize) {
 		printf("The resource could not be read!\n");
 		free(loBuffer);
