+++ apps/daesop/utils.cpp
@@ -83,8 +83,8 @@ int stringEndsWith(char *aFullString, const char *aEndString) {
 		printf("stringEndsWith(): at least one parameter is NULL!\n");
 		return (loResult);
 	}
-	loFullStringLength = strlen(aFullString);
-	loEndStringLength = strlen(aEndString);
+	loFullStringLength = static_cast<int>(strlen(aFullString));
+	loEndStringLength = static_cast<int>(strlen(aEndString));
 	if (loFullStringLength < loEndStringLength) {
 		// the second string is longer
 		return (loResult);
@@ -116,7 +116,7 @@ int copyFile(FILE *aSourceFile, char *aNewFileName) {
 
 	// copy the aResFile file into aNewFile
 	fseek(aSourceFile, 0, SEEK_SET);
-	while ((loReadSize = fread(&loBuffer, 1, MAX_COPY_BUFFER, aSourceFile)) != 0) {
+	while ((loReadSize = static_cast<unsigned int>(fread(&loBuffer, 1, MAX_COPY_BUFFER, aSourceFile))) != 0) {
 		if (fwrite(loBuffer, 1, loReadSize, loNewFile) != loReadSize) {
 			printf("Unable to write to the file: %s\n", aNewFileName);
 			fclose(loNewFile);
